refactors file structure for app handlers
