+++ internal/handlers/remoteEvents/guildJoinEvent/join.go
