+++ internal/handlers/remoteEvents/guildRemoveEvent/guildRemove.go
