+++ internal/handlers/remoteEvents/readyEvent/assets/defaultContent/default.txt
