+++ internal/handlers/remoteEvents/readyEvent/assets/defaultContent/legends.txt
