+++ internal/handlers/remoteEvents/readyEvent/assets/defaultContent/music-info.txt
