+++ internal/handlers/remoteEvents/readyEvent/assets/defaultContent/server-rules.txt
