+++ internal/handlers/remoteEvents/readyEvent/assets/defaultContent/staff-rules.txt
