+++ internal/handlers/remoteEvents/readyEvent/load.go
