+++ internal/handlers/slashEvents/commands/server/help.go
