+++ internal/handlers/slashEvents/register.go
@@ -6,7 +6,7 @@ import (
 
 	globalConfiguration "github.com/RazvanBerbece/Aztebot/internal/globals/configuration"
 	globalState "github.com/RazvanBerbece/Aztebot/internal/globals/state"
-	"github.com/RazvanBerbece/Aztebot/internal/handlers/slashCommandEvent/commands"
+	"github.com/RazvanBerbece/Aztebot/internal/handlers/slashEvents/commands"
 	"github.com/RazvanBerbece/Aztebot/pkg/shared/utils"
 	"github.com/bwmarrin/discordgo"
 )
