+++ internal/handlers/slashEvents/slashHandler.go
@@ -9,8 +9,8 @@ import (
 	globalConfiguration "github.com/RazvanBerbece/Aztebot/internal/globals/configuration"
 	globalMessaging "github.com/RazvanBerbece/Aztebot/internal/globals/messaging"
 	globalRepositories "github.com/RazvanBerbece/Aztebot/internal/globals/repositories"
-	actionEvent "github.com/RazvanBerbece/Aztebot/internal/handlers/actionEvents"
-	"github.com/RazvanBerbece/Aztebot/internal/handlers/slashCommandEvent/commands"
+	actionEvent "github.com/RazvanBerbece/Aztebot/internal/handlers/remoteEvents/actionEvents"
+	"github.com/RazvanBerbece/Aztebot/internal/handlers/slashEvents/commands"
 	"github.com/RazvanBerbece/Aztebot/pkg/shared/utils"
 	"github.com/bwmarrin/discordgo"
 )
